lesson_generator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lesson-generator"
version = "0.1.0"
description = "AI-powered lesson generator for programming courses"
readme = "README.md"
authors = [{name = "LibertyQuinzel", email = "example@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["education", "programming", "lesson", "generator", "ai"]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "jinja2>=3.1.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "structlog>=23.1.0",
    "aiohttp>=3.8.0",
    "tenacity>=8.2.0",
    "jsonschema>=4.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "pylint>=2.17.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "pre-commit>=3.3.0",
    "memory-profiler>=0.61.0",
]
docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[project.urls]
Homepage = "https://github.com/LibertyQuinzel/lesson-generator"
Repository = "https://github.com/LibertyQuinzel/lesson-generator"
Documentation = "https://lesson-generator.readthedocs.io"
Issues = "https://github.com/LibertyQuinzel/lesson-generator/issues"

[project.scripts]
lesson-generator = "lesson_generator.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=lesson_generator",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=95",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, external deps)",
    "e2e: End-to-end tests (slowest, full system)",
    "performance: Performance and load tests",
]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pylint.messages_control]
disable = [
    "missing-docstring",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-instance-attributes",
    "duplicate-code",
]

[tool.pylint.format]
max-line-length = "88"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.coverage.run]
# Use workspace-relative paths so coverage patterns match regardless of CWD
relative_files = true
source = [
    "lesson_generator/src/lesson_generator",
]
omit = [
    # Test files
    "*/tests/*",
    "*/test_*",
    # Exclude external API integration and heavy orchestrators from coverage for now
    "*/src/lesson_generator/content/openai_generator.py",
    "*/src/lesson_generator/core/template_extractor.py",
    "*/src/lesson_generator/core/generator.py",
    "*/src/lesson_generator/content/__init__.py",
    "*/src/lesson_generator/cli/main.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*Protocol\\):",
    "@(abc\\.)?abstractmethod",
]